pyproject.toml
[tool.poetry]
name = "telegram-selenium"
version = "0.1.0"
description = ""
authors = ["Davies <Davies.araujo@fazobetai.com>"]
readme = "README.md"
packages = [{include = "telegram_selenium"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
selenium = "^4.9.1"
python-dotenv = "^1.0.0"
splinter = "^0.19.0"
pandas = "^2.0.1"
sqlalchemy = "^2.0.15"
pymysql = "^1.0.3"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
webdriver-manager = "^3.8.6"
ipykernel = "^6.23.1"
pyinstaller = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
